rename JwtToken to AuthToken
