+++ lib/msal-common/src/client/SilentFlowClient.ts
@@ -12,7 +12,7 @@ import { IdTokenEntity } from "../cache/entities/IdTokenEntity";
 import { AccessTokenEntity } from "../cache/entities/AccessTokenEntity";
 import { RefreshTokenEntity } from "../cache/entities/RefreshTokenEntity";
 import { ScopeSet } from "../request/ScopeSet";
-import { JwtToken } from "../account/JwtToken";
+import { AuthToken } from "../account/AuthToken";
 import { TimeUtils } from "../utils/TimeUtils";
 import { RefreshTokenRequest } from "../request/RefreshTokenRequest";
 import { RefreshTokenClient } from "./RefreshTokenClient";
@@ -77,7 +77,7 @@ export class SilentFlowClient extends BaseClient {
         // Return tokens from cache
         this.config.serverTelemetryManager.incrementCacheHits();
         const cachedIdToken = this.readIdTokenFromCache(homeAccountId, environment, cachedAccount.realm);
-        const idTokenObj = new JwtToken(cachedIdToken.secret, this.config.cryptoInterface);
+        const idTokenObj = new AuthToken(cachedIdToken.secret, this.config.cryptoInterface);
 
         return await ResponseHandler.generateAuthenticationResult(
             this.cryptoUtils,
