+++ lib/msal-common/src/crypto/PopTokenGenerator.ts
@@ -1,5 +1,5 @@
 import { ICrypto } from "./ICrypto";
-import { JwtToken } from "../account/JwtToken";
+import { AuthToken } from "../account/AuthToken";
 import { TokenClaims } from "../account/TokenClaims";
 import { TimeUtils } from "../utils/TimeUtils";
 import { UrlString } from "../url/UrlString";
@@ -40,7 +40,7 @@ export class PopTokenGenerator {
     }
 
     async signPopToken(accessToken: string, resourceRequestMethod: string, resourceRequestUri: string): Promise<string> {
-        const tokenClaims: TokenClaims = JwtToken.extractTokenClaims(accessToken, this.cryptoUtils);
+        const tokenClaims: TokenClaims = AuthToken.extractTokenClaims(accessToken, this.cryptoUtils);
         const resourceUrlString: UrlString = new UrlString(resourceRequestUri);
         const resourceUrlComponents: IUri = resourceUrlString.getUrlComponents();
         return await this.cryptoUtils.signJwt({
