+++ lib/msal-common/src/index.ts
@@ -7,7 +7,7 @@ export { AuthOptions, SystemOptions, LoggerOptions, DEFAULT_SYSTEM_OPTIONS } fro
 export { ClientConfiguration } from "./config/ClientConfiguration";
 // Account
 export { AccountInfo } from "./account/AccountInfo";
-export { JwtToken } from "./account/JwtToken";
+export { AuthToken as JwtToken } from "./account/AuthToken";
 export { TokenClaims } from "./account/TokenClaims";
 // Authority
 export { Authority } from "./authority/Authority";
